Refactor with AE Support
Refactor the base we got from LibFire, and add support for building
for AE.

A few functions got lost in the process:

* Race::GetSkillBonus
* ObjectREFR::GetPermanentActorValue

In the past it seems like CommonLib had a function to look up
actor values by name. This functionality is gone in recent
versions, thus if we want to get these functions back we'll
need to re-implement them ourselves.